#6
Fixed issue with Master being able to delete bids. Master can now delete
any bid at any phase.

Added protection against attempting to edit a shipment/delete a shipment
or bid while the shipment is frozen if user is not MASTERSERVER.
Shippers also cannot accept bids on frozen shipments now

Shippers can now see their own frozen shipments (only their own, not
all) but not interact with them.

KNOWN ISSUES:
Several redirects go  back to the wrong part of the shipments screen and
need to be repaired, including clicking "no" from any of the
confirmation screen. A few of the shipments screens of no back button or
back buttons that redirect to the wrong page.

MasterServer cannot edit bids or shipments. Clicking the buttons to do
so will cause an error.

The "view bids" and "view details" buttons  are on different sides of
the screens in different menus.

Carriers can edit bids on already completed auctions. This was built
this way from the start, so I question if this is intentional?