#7
Master now has the ability to force complete shipments, however, there a
few issues with it. It fails if a bid more than ten digits long is
placed on the shipment when trying to force close. Method also lacks
proper error logging. Will leave issue open and branch unmerged until
this is fixed.