close #3
New feature, when shippers create a shipment it is now placed in
"PENDING" instead of available. Pending shipments cannot be bid on, and
will not show up in the auctioning screen. Pending shipments should
always have no bids.

SHIPPERS and MASTERSERVER can push shipments from pending into auction,
marking them as "AVALIABLE SHIPMENT" and allowing them to be bid on.