#58
Backup system is now up and working. Admin can load the backup and
restore the database, and a new backup is created every 20 minutes.
Backup loading works very well, refreshes any changes that happened
within the last twenty minutes.

To do this is requires bundling in another binary, the mysql binary,
which is included under src/main/resources/static/binaries. These will
have to be included as a seperate bin folder when extracting the project
as a war file

Once I add the capabilites for more than one backup at a time to be
saved (and have them labeled by data/time) and add finishing touches to
the system I will close #58