close #41
Admin user can now properly delete shippers and carriers, and when they
are deleted, all of their dependent sub-data is deleted (shipments,
bids, and notifications). This is the first major step to resolving any
"ripple effects" that come from deleting accounts.

All account types can now easily be deleted and notifications are sent
out to any users affected by the deletion. Vehicles, contacts, and
drivers are still left over in the database but this should be easy to
fix/clean up.