Close#77
Carrier contact upload through excel now working
along with input validation and error handling.
Sample contact .xlsx file added for the template.
Also made it so that if the user clicks the upload button without
selecting the file, they will be prompted to upload the file.
I added that on all of the already created upload files as well.