#70
Cleaned up the select carrier page, added a frame for error messages

#79 #71

The problem is related to stray "." symbols in the
windows.locations.href command attached to the "back" and "add" buttons.
I have fixed the error for vehciles types, solution should be portable
to other parts of the application.