close #70
Shippers can now bypass the auction and send a request to pick up a
shipment directly to a carrier. Carriers can then either Accept or Deny
the request to take that shipment. If accepted, it is moved to accepted
status and marked with the correct price. If denied, the shipment is
moved back to pending.

This should be integrated with #76 to have emails sent out about the
interactions with this standalone system.