close #76
Notification system is now integrated into emailing. You can chose if a
notification should be also emailed by either putting true or false at
the end of the addNotification function. Currently emails are sent out
to shippers and carriers for direct assignment acceptance and denials,
but this system is modular and can be used to send an email for any
notification.

Also started rebuilding all of the static methods inside
notificationcontroller to be part of a service class rather than being
static