#109
Built a function for reading a freight rate table and fetching the
correct shipment price from it. All that should be left is plugging
everything together and the feature will be complete.

The freight rate table works on the basis of weight and commodity class.
Shippers can assign different prices to shipments based on these two
factors, as well as have different tables for different carriers. A
template excel file will be provided, as well as additional information
in the user manual.