#109
The freight rate table is now working. Built javascript functions to
pull the freight rate table values onto the direct assignment page when
the user clicks a button. This way, the user can easily chose to
manually assign a price or use the freight rate table.

A few minor finishing touches are still needed. Better error messaging
to the user, and a way for users to add a default price column for
shipments of weights over the defined weight limits.