close #81
I have fixed the header into position above the content div. Now the
sidebar will no longer become unbound if the size of the content div
becomes too large. The header and navbar will now always be fixed in
place and the scrollwheel only moves the content area of the page. This
improves the overall look and feel of the user interface. (Header could
probably be shrunk a bit, and the boarder intensity of the navbar
reduced to make it appear more "behind" the header)