close #110
Added some CSS to shrink the sidebar elements and the header when the
user shrinks the page window/is on a smaller monitor. Should ensure that
the entire sidebar is always visible.

#134
Carrier sidebar now has dropdown menus for easier access.

#135
All carrier related pages now use the fragment sidebar instead of the
copy-pasted one. Should only be a few left over in the whole
application.