close #135
The sidebar is now one single file in the application, that is
distributed among all pages as a thymeleaf fragment, rather than being
copy pasted. The same has been done with core pageload javascript. This
removes nearaly 10,000 redundant lines of code from the program, and
makes it far, far easier for future programmers to make changes to the
sidebar.

close #134

All redundant selection menus have been replaced with dropdowns on the
sidebar. This includes mainly the shipments pages and the user pages for
admins.