close #111
Admin page now has dropdowns instead of redundant pages. Toggle buttons
have been removed and placed in edit screens instead of clogging up the
tables. Footer has been rooted to the bottom of the page To prevent if
from dissapearing if the user list is too large. General cleaning up of
the HTML. The add-shipper and add-carrier button still need to be worked
on. Opening a seperate issue for that. Visually, admin page looks pretty
nice now.

In addition, the sidebar buttons now properly lights up the right pages
no matter which sub page you are on.